requirements.txt
django
djangorestframework
requests
re